+++ Packages/Tracking/Examples~/XR Examples/Example Assets/Misc and Shared/Scripts/SimpleGrabBallFeedback.cs
@@ -6,7 +6,7 @@
  * between Ultraleap and you, your company or other organization.             *
  ******************************************************************************/
 
-using Leap.Unity.Interaction;
+using Leap.Unity.PhysicalHands;
 using UnityEngine;
 
 namespace Leap.InteractionEngine.Examples
@@ -56,10 +56,10 @@ namespace Leap.InteractionEngine.Examples
 
             defaultMesh.transform.rotation = LookAtRotationParallelToHorizon(defaultMesh.transform.position, Camera.main.transform.position);
 
-            bool expanded = grabBall.grabBallInteractionBehaviour.closestHoveringControllerDistance < distanceToScaleGrabBall;
+            bool expanded = grabBall.ClosestHandDistance < distanceToScaleGrabBall;
             defaultMesh.transform.localScale = Vector3.Lerp(defaultMesh.transform.localScale, (expanded) ? expandedScale : minimisedScale, Time.deltaTime * lerpTime);
 
-            if (grabBall.grabBallRestrictionStatus.IsRestricted && grabBall.grabBallInteractionBehaviour.isGrasped)
+            if (grabBall.grabBallRestrictionStatus.IsRestricted && grabBall.IsGrabbed)
             {
                 ghostedMesh.gameObject.SetActive(true);
                 ghostedMesh.transform.position = grabBall.grabBallPose.position;